# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Playwright & Selenium Learning Platform API
  description: |
    The Playwright & Selenium Learning Platform API allows developers to integrate learning features,
    access user data, manage webhooks, and build custom integrations.

    ## Authentication

    The API supports two authentication methods:

    ### API Keys
    Include your API key in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_KEY
    ```

    ### OAuth 2.0
    Use OAuth 2.0 for user-authorized access:
    ```
    Authorization: Bearer YOUR_ACCESS_TOKEN
    ```

    ## Rate Limiting

    API requests are rate-limited based on your API key configuration (default: 1000 requests/hour).
    Rate limit information is included in response headers:
    - `X-RateLimit-Limit`: Total requests allowed per hour
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Time when the limit resets

    ## Scopes

    API keys and OAuth tokens require specific scopes for different operations:
    - `users:read` - Read user information
    - `users:write` - Update user information
    - `lessons:read` - Access lesson content
    - `lessons:write` - Create/update lessons
    - `quizzes:read` - Access quizzes
    - `quizzes:write` - Create/update quizzes
    - `webhooks:read` - View webhooks
    - `webhooks:write` - Manage webhooks
    - `api_keys:read` - View API keys
    - `api_keys:write` - Manage API keys

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
paths:
...
  /webhooks:
    get:
      tags:
        - Webhooks
      summary: List webhooks
      description: Get all webhooks for the authenticated user
      operationId: listWebhooks
      security:
        - ApiKeyAuth: []
        - OAuth2: [webhooks:read]
      responses:
        '200':
          description: Webhooks retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer

    post:
      tags:
        - Webhooks
      summary: Create webhook
      description: Create a new webhook
      operationId: createWebhook
      security:
        - ApiKeyAuth: []
        - OAuth2: [webhooks:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      webhook:
                        $ref: '#/components/schemas/Webhook'
                      secret:
                        type: string
                        description: Webhook secret for signature verification
  /webhooks/{webhookId}:
    get:
      tags:
        - Webhooks
      summary: Get webhook
      description: Get a specific webhook by ID
      operationId: getWebhook
      security:
        - ApiKeyAuth: []
        - OAuth2: [webhooks:read]
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: Webhook retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Webhook'

    patch:
      tags:
        - Webhooks
      summary: Update webhook
      description: Update an existing webhook
      operationId: updateWebhook
      security:
        - ApiKeyAuth: []
        - OAuth2: [webhooks:write]
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/Webhook'

    delete:
      tags:
        - Webhooks
      summary: Delete webhook
      description: Delete a webhook
      operationId: deleteWebhook
      security:
        - ApiKeyAuth: []
        - OAuth2: [webhooks:write]
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          $ref: '#/components/responses/Success'
  /webhooks/{webhookId}/test:
    post:
      tags:
        - Webhooks
      summary: Test webhook
      description: Send a test event to the webhook
      operationId: testWebhook
      security:
        - ApiKeyAuth: []
        - OAuth2: [webhooks:write]
      parameters:
        - $ref: '#/components/parameters/WebhookId'
      responses:
        '200':
          description: Test webhook delivered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      success:
                        type: boolean
                      status:
                        type: integer
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: Use your API key as a bearer token

    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://api.playwright-learning.com/oauth/authorize
          tokenUrl: https://api.playwright-learning.com/v1/oauth/token
          scopes:
            users:read: Read user information
            users:write: Update user information
            lessons:read: Access lesson content
            lessons:write: Create/update lessons
            quizzes:read: Access quizzes
            quizzes:write: Create/update quizzes
            webhooks:read: View webhooks
            webhooks:write: Manage webhooks
            api_keys:read: View API keys
            api_keys:write: Manage API keys
  parameters:
...
    WebhookId:
      name: webhookId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
            enum:
              - user.created
              - user.updated
              - lesson.completed
              - quiz.passed
              - quiz.failed
              - exercise.completed
              - achievement.unlocked
              - review.completed
              - progress.milestone
              - subscription.changed
        isActive:
          type: boolean
        description:
          type: string
        successCount:
          type: integer
        failureCount:
          type: integer
        lastTriggeredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    CreateWebhookRequest:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        description:
          type: string
        headers:
          type: object
        retryPolicy:
          type: object
          properties:
            maxAttempts:
              type: integer
              minimum: 1
              maximum: 5
            backoffMultiplier:
              type: integer
              minimum: 1
    UpdateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            type: string
        description:
          type: string
        isActive:
          type: boolean
        headers:
          type: object
...
  responses:
    Success:
      description: Operation successful
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              meta:
                type: object
                properties:
                  message:
                    type: string
...